Changed the consume_error trait to exit from execution. Added consume_errors trait.
The consume error trait used to be (Result<T,E>) -> Result(T, ()) and just printed the error.
Now it is (Result<T,E>) -> T and if an error occured it prints it and exits.
There is also a variant called consume_errors that does the same but is defined for
(Vec<Result<T,E>>) -> Vec<T>.